Further test consolidation
